requirements.txt
# Core application dependencies
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
python-dotenv==1.0.0

# Data visualization
altair==5.1.2
plotly==5.17.0

# LLM and vector search
langchain==0.0.335
langchain-openai==0.0.2
chromadb==0.4.15
faiss-cpu==1.7.4

# Database connectors
duckdb==0.9.2
snowflake-connector-python==3.5.0
google-cloud-bigquery==3.13.0

# SQL parsing and validation
sqlglot==19.3.0
sqlparse==0.4.4

# dbt
dbt-core==1.6.8
dbt-duckdb==1.6.2
dbt-snowflake==1.6.2
dbt-bigquery==1.6.8

# Development and testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
ruff==0.1.6
pre-commit==3.5.0

# Utilities
pyyaml==6.0.1
click==8.1.7
requests==2.31.0